Non-used files removed
