--- GccMatrixWS2812 V3/GccMatrixWS2812 V3/src/cambiarJugador.c
@@ -1,6 +1,0 @@
-/*
- * cambiarJugador.c
- *
- * Created: 1/12/2022 16:29:23
- *  Author: NinoM
- */
